@@ -21,8 +21,8 @@
     "@svgr/webpack": "^6.5.1",
     "@types/js-cookie": "^3.0.2",
     "@types/node": "^18.11.18",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
     "@types/react-router-dom": "^5.3.3",
     "@types/validator": "^13.7.10",
     "babel-loader": "^9.1.2",
@@ -37,22 +37,22 @@
     "source-map-loader": "^4.0.1",
     "style-loader": "^3.3.1",
     "ts-loader": "^9.4.2",
-    "typescript": "^4.9.4",
+    "typescript": "^5.0.2",
     "webpack": "^5.75.0",
     "webpack-cli": "^5.0.1",
     "webpack-dev-server": "^4.11.1"
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.1",
-    "@twilio-paste/core": "^17.2.0",
-    "@twilio-paste/icons": "^9.4.2",
+    "@twilio-paste/core": "^18.1.0",
+    "@twilio-paste/icons": "^10.0.0",
     "axios": "^1.2.3",
     "js-cookie": "^3.0.1",
     "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-redux": "^8.0.5",
-    "react-router-dom": "^5.3.3",
+    "react-router-dom": "^6.9.0",
     "validator": "^13.7.0"
   }
 }
