@@ -1,6 +1,4 @@
-import { Text } from '@twilio-paste/core/text';
-import { Alert } from '@twilio-paste/core/alert';
-import { Anchor } from '@twilio-paste/core/anchor';
+import { BasicErrorAlert, ContactSupportLink } from '../../../components/BasicErrorAlert';
 
 export interface SignupErrorAlertProps {
   title?: string;
@@ -9,11 +7,7 @@ export interface SignupErrorAlertProps {
 export const SignupErrorAlert = ({
   title = 'An unknown error has occurred.',
 }: SignupErrorAlertProps) => (
-  <Alert variant="error">
-    <Text as="span">
-      <strong>{title}</strong> Please try again or{' '}
-      <Anchor href="/support/contact">contact support</Anchor> if the problem
-      continues.
-    </Text>
-  </Alert>
+  <BasicErrorAlert>
+    <strong>{title}</strong> Please try again or <ContactSupportLink /> if the problem continues.
+  </BasicErrorAlert>
 );
